mark1/requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
llama-cpp-python




pip install streamlit==1.34.0
pip install sentence-transformers==2.2.2
pip install PyPDF2==3.0.1
pip install scikit-learn==1.4.2
pip install llama-cpp-python==0.2.74
